// client/src/lib/queries/queryKeys.ts
export enum QUERY_KEYS {
  // AUTH KEYS
  CREATE_USER_ACCOUNT = "createUserAccount",

  // USER KEYS
  GET_CURRENT_USER = "getCurrentUser",
  GET_API_KEYS = "getApiKeys",
  GET_PAYMENTS = "getPayments",
  GET_USERS = "getUsers",
  GET_USER_BY_ID = "getUserById",
  GET_ALL_PAID_USER = "allPaidUser",
  GET_ALL_USERS_BY_DATE = "allUsersByDate",
  GET_JOBS_IN_DB = "getjobsindb"

}